getValue didn't update the latest node after walking through
the path and may regard the static route as the param type,
but the params cap was preallocated and cause out of range.

For Example:

the route was "/a/b/:c/d" and we send the request path "/a/b/c/d1",
the params cap was 1 but the `d` would also be regarded as param then
would be out of range.