@maximerety 2021-08-23
Hi @git-hulk,

We encountered the same issue (while trying to upgrade to gin 1.7.3) but I wasn't sure how to tackle it.

So thank you for the fix and the explanations!

I guess you could add a test case to `gin_integration_test.go` too, since it seems to mirror all tests from `tree_test.go`, e.g.:

```diff
diff --git a/gin_integration_test.go b/gin_integration_test.go
index a4149de..cb05f3a 100644
--- a/gin_integration_test.go
+++ b/gin_integration_test.go
@@ -463,4 +463,5 @@ func TestTreeRunDynamicRouting(t *testing.T) {
        testRequest(t, ts.URL+"/a/dd", "404 Not Found")
        testRequest(t, ts.URL+"/addr/dd/aa", "404 Not Found")
        testRequest(t, ts.URL+"/something/secondthing/121", "404 Not Found")
+       testRequest(t, ts.URL+"/cc/dd/ee/ff/gg/hh1", "404 Not Found")
 }
```

I hope this fix will be merged/released soon so that we'll be able to upgrade gin beyond 1.7.2.