@git-hulk 2021-08-24
@maximerety Thanks for the suggestion,  we are also looking forward to fixing this issue soon since it would cause the request panic.