@appleboy 2021-08-25
@git-hulk Thanks for your effort. Can you also provide the benchmark result like this comment: https://github.com/gin-gonic/gin/pull/2767#issuecomment-871934041

Hi @qm012 @rw-access, please also take a look if you have free time.