@git-hulk 2021-08-29
@appleboy ok, I’ll post the benchmark result later, sorry for missing your previous comment.