@git-hulk 2021-08-29
The benchmarks:
master https://app.travis-ci.com/github/git-hulk/go-http-routing-benchmark/builds/236444329
this PR https://app.travis-ci.com/github/git-hulk/go-http-routing-benchmark/builds/236442891

The benchstat report:

```
BenchmarkGin_Param        	58612060	        61.18 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	31930453	       117.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	11869882	       302.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	31239873	       117.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	44474041	        80.83 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	27028000	       135.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	27133978	       132.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	  135966	     26346 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	59091340	        60.47 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	39117842	        92.22 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	31236037	       115.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 3110154	      1176 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	61985156	        59.47 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	55725441	        65.23 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	44601469	        80.90 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1842114	      2007 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	  191962	     18623 ns/op	       0 B/op	       0 allocs/op

```