@appleboy 2021-08-29
benchmark result:

```
$ benchstat a.txt b.txt 
name              old time/op    new time/op    delta
Gin_Param           57.9ns ± 1%    60.5ns ± 5%   +4.46%  (p=0.000 n=9+10)
Gin_Param5           102ns ± 3%     115ns ± 3%  +13.64%  (p=0.000 n=10+10)
Gin_Param20          254ns ± 2%     304ns ± 1%  +19.90%  (p=0.000 n=10+10)
Gin_ParamWrite       117ns ± 2%     116ns ± 4%     ~     (p=0.150 n=9+10)
Gin_GithubStatic    72.0ns ± 0%    81.0ns ± 1%  +12.48%  (p=0.000 n=10+10)
Gin_GithubParam      122ns ± 1%     133ns ± 0%   +8.55%  (p=0.000 n=10+9)
Gin_GithubAll       25.3µs ± 0%    26.3µs ± 0%   +4.14%  (p=0.000 n=9+10)
Gin_GPlusStatic     57.9ns ± 0%    60.7ns ± 0%   +4.80%  (p=0.000 n=10+10)
Gin_GPlusParam      76.4ns ± 0%    91.9ns ± 1%  +20.31%  (p=0.000 n=10+10)
Gin_GPlus2Params    95.7ns ± 0%   116.1ns ± 2%  +21.30%  (p=0.000 n=9+9)
Gin_GPlusAll        1.07µs ± 1%    1.17µs ± 1%   +9.03%  (p=0.000 n=9+10)
Gin_ParseStatic     57.5ns ± 0%    58.0ns ± 1%   +0.87%  (p=0.000 n=10+8)
Gin_ParseParam      63.4ns ± 0%    64.9ns ± 1%   +2.36%  (p=0.000 n=10+10)
Gin_Parse2Params    76.2ns ± 0%    80.9ns ± 0%   +6.11%  (p=0.000 n=10+10)
Gin_ParseAll        1.82µs ± 1%    1.98µs ± 1%   +8.83%  (p=0.000 n=10+10)
Gin_StaticAll       18.3µs ± 1%    18.6µs ± 0%   +1.81%  (p=0.000 n=10+10)
```