@git-hulk 2021-08-29
@appleboy so should I worry about the performance degrade?