@appleboy 2021-08-29
@git-hulk Yes, we need to care about the performance of the router. @qm012 Do you have any comments about the benchmark result?