@qm012 2021-08-30
> @git-hulk Yes, we need to care about the performance of the router. @qm012 Do you have any comments about the benchmark result?

hi @appleboy Migration is done because it affects route matching performance and makes multiple circular calls.If fixing this bug results in such a large performance difference, I think it is not worth the cost, so we can avoid this bug in route definition for the time being until we have a better solution.

Test report after migration https://github.com/gin-gonic/gin/pull/2796#issuecomment-885981240

Before the migration
![image](https://user-images.githubusercontent.com/67568757/131273300-622bf8d1-f702-4019-94dc-5454e39d9ee9.png)
After the migration
![image](https://user-images.githubusercontent.com/67568757/131273356-e9662dfd-aed8-448a-945d-ff419521aa63.png)