@qm012 2021-08-30
> @qm012 I don't think we should NOT sacrifice the correctness to promise the performance, right?

They are compatible, so we need to talk about them.
@appleboy @thinkerou @rw-access What do you think