@qm012 2021-08-30
> @qm012 We should NOT sacrifice the correctness to promise the performance, right?

There are many ways to solve a problem.
Like this https://github.com/gin-gonic/gin/pull/2767#issuecomment-871053458 can fix problem but the test report is not good, we can optimize until we get a better test report https://github.com/gin-gonic/gin/pull/2767#issuecomment-876083905