@averageflow 2021-09-27
Issue still exists, thus i am frozen at gin 1.7.2... see #2878